Add get endpoints and resources functions to MBEDCloudSession
One function gets all registered devices, the other function
reads endpoint resource.

One catch: reading an endpoint resource is asynchronous, so all
that is returned is an async-id. You'll only get the actual
value from a resource either from the longpoll or the webhook
callback. You will have to keep track of the async-id somehow
so you know what the value you receive is for..

Signed-off-by: mbanders <michael.anderson@arm.com>